Splitted Admin and USer into different modules
